requirements.txt
requests>=2.31.0
# Used by ADSBDB enrichment (urllib) so macOS/Python.org installs verify TLS like requests does.
certifi>=2023.5.7
pytest>=8.0.0
python-dotenv>=1.0.1
